+++ src/main/java/me/marin/lockout/lockout/goal/rendering/texture/TextureExtractor.java
@@ -8,7 +8,7 @@ public interface TextureExtractor {
      * Renders/displays any custom texture. This method is called every client tick, meaning that the texture can change.
      * @param tick Number of client ticks passed. This number increases every client tick.
      */
-    void extract(GuiGraphicsExtractor extractor, Font font, int x, int y, int width, int height, int tick);
+    void extract(GuiGraphicsExtractor extractor, Font font, int x, int y, int width, int height, long tick);
 
     default void withScale(GuiGraphicsExtractor extractor, int x, int y, float scale, Runnable function) {
         extractor.pose().pushMatrix();
